iOS: снять неоднозначность оператора в AetherMediaViewer — падало на Xcode 26.1
Третья и последняя причина, по которой workflow iOS build не проходил.
Color.opacity принимает Double, dismissProgress объявлен CGFloat, и выражение
"1 - dismissProgress * 0.7" на Xcode 26.1.1 (его ставит CI) считается
неоднозначным: компилятор не выбирает перегрузку "-". На Xcode 26.6, который
стоит локально, вывод типов справляется сам — поэтому локальная сборка была
зелёной, а CI падал. Приведение к Double снимает неоднозначность и ничего не
меняет по смыслу.

Файл к правкам аудита отношения не имеет: последний раз менялся в c986b9f.
Проверено локальной сборкой — BUILD SUCCEEDED.

+++ ios/AETHER/Features/Chat/AetherMediaViewer.swift
@@ -91,7 +91,12 @@ struct AetherMediaViewer: View {
 
     var body: some View {
         ZStack {
-            Color.black.opacity(1 - dismissProgress * 0.7).ignoresSafeArea()
+            // Double(...) не для красоты: opacity принимает Double, а
+            // dismissProgress — CGFloat, и Swift 26.1 объявляет «1 - x * 0.7»
+            // неоднозначным (какой из перегруженных «-» брать). На 26.6 вывод
+            // типов справляется сам, поэтому локально ошибки не видно, а сборка
+            // в CI падает. Приводим явно.
+            Color.black.opacity(Double(1 - dismissProgress * 0.7)).ignoresSafeArea()
 
             pages
                 .offset(y: dragY)
